New follow movement generator for CMaNGOS (transitional)
Replaces existing follow movement generator with a mew one, which more closely resembles original game.

New logic is much more responsive to target's movement.

A few features were not finished for this patch and will need more work in the future.

The list of changes includes, but not limited to:
- Less intensive position polling update interval
- Fake prediction of unit's movement while in motion (unfinished)
- Speed boost for pets allowing pets to catch up with master when out of combat
- Additional facing adjustment tick to closely simulate original game
- Auto-unstuck for permanent pet followers